add more validation to delete product route

+++ server/src/controllers/product.js
@@ -1,4 +1,3 @@
-import Axios from "axios";
 import mongoose from "mongoose";
 import { Product } from "../models/product.js";
 import { BadRequest, Forbidden, NotFound } from "../utils/error.js";
@@ -33,6 +32,7 @@ export class ProductController {
       next(ex);
     }
   }
+
   static async createProduct(req, res) {
     const { userId } = req.user || {};
 
@@ -61,16 +61,17 @@ export class ProductController {
     });
   }
 
-  static async deleteProduct(req, res, next) {
-    try {
-      const deletedProduct = await Product.findById(req.params.id);
-      if (!deletedProduct) throw new NotFound("Product not found");
-      if (req.user.userId !== deletedProduct.owner.toHexString())
-        throw new Forbidden("Product does not belongs to you");
-      await deletedProduct.deleteOne();
-      res.json({ message: "successful", deletedProduct });
-    } catch (error) {
-      next(error);
-    }
+  static async deleteProduct(req, res) {
+    const { id } = req.params;
+    if (!mongoose.Types.ObjectId.isValid(id))
+      throw new BadRequest({ id: `Invalid product id '${id}'` });
+
+    const deletedProduct = await Product.findById(id);
+    if (!deletedProduct) throw new NotFound("Product not found");
+
+    if (req.user.userId !== deletedProduct.owner.toHexString())
+      throw new Forbidden("Product does not belongs to you");
+    await deletedProduct.deleteOne();
+    res.json({ message: "successful", deletedProduct });
   }
 }
